requirements.txt
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
greenlet==3.1.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==17.0.0
Pygments==2.18.0
pyspark==3.5.2
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
six==1.16.0
SQLAlchemy==2.0.35
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
python-dotenv==1.0.1
wcwidth==0.2.13
kafka-python-ng==2.2.2